requirements.txt
certifi==2016.2.28
mido==1.2.8
numpy>=1.13.1
-e git+https://github.com/generalelectrix/pyenttec.git@4d0ba697f992c8c13f614a35b31458e39aa8a1ee#egg=pyenttec
pyserial==3.4
# -e git+https://github.com/c0z3n/pySimpleDMX.git@13f9d5b5c7e54500f097e700eb3289df87501306#egg=pysimpledmx
python-dateutil==2.6.1
python-rtmidi==1.1.0
pytz==2017.2
six==1.10.0
wincertstore==0.2
pyyaml==5.1.1
scipy==1.3.0
serial==0.0.97
python-osc==1.7.2